+++ daiv_sandbox/config.py
@@ -1,7 +1,7 @@
 import warnings
 from typing import Literal
 
-from pydantic import Field, HttpUrl, SecretStr
+from pydantic import HttpUrl, SecretStr
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 warnings.filterwarnings(
@@ -32,8 +32,9 @@ class Settings(BaseSettings):
 
     # Execution
     RUNTIME: Literal["runc", "runsc"] = "runc"
-    KEEP_TEMPLATE: bool = False
-    MAX_EXECUTION_TIME: int = Field(default=600, description="Maximum execution time in seconds")
+
+    # Git
+    GIT_IMAGE: str = "alpine/git"
 
 
 settings = Settings()  # type: ignore